BgeeProperties in bgee-core + bgee-webapp
I took the most recent BgeeProperties behavior from Bgee-webapp and
implemented it in bgee-core. I turned the bgee-webapp BgeeProperties
into a subclass. BgeeWebappProperties in bgee-webapp extends now
BgeeProperties in bgee-core
I adapted all the tests. It needed a change in the bgee-core pom.xml in
order to run the test independently ( 1 thread for each tests )
I duplicated the tests, so some parts are run twice, once in bgee-core
and once in bgee web app… it might be a bit redundant but it ensures
that the dependency between the module is 100% correctly handled.